Update Packages-Desktop
Mate system monitor isn't included in the live cd and it's really useful. It also has a tab `About` which many users (and youtubers) use it to see the current MATE version before they install it.

I think mate-tweak should be included in the full installation. It is useful, especially for the new users.